Add Orcsoberfest areas that do not overlap with Winterlight Solstice island
Orcsoberfest island is a big problem, since it’s in the exact same map location as the Winterlight Solstice island. That is, our map data cannot simultaneously contain both areas in their fully mapped form.

For this reason, we’re adding only the (parts of the) areas that do not overlap to the map data.

The Great Schnitzel boss room: https://tibiamaps.io/map#33713,31116,9:1

Ref. #83.