Improve marker label consistency
We tend to use sentence case for labels instead of title case. In other words, we only capitalize the start of the label, and any names of NPCs or places in there, but nothing else.

Also use “ominous cauldron” instead of “omnious cauldron” as that’s an obvious typo in-game that CipSoft should really fix.